refactor: split orchestration.py into disks/context.py + run/
- managed_mount context manager moves to disks/context.py
  (only depends on disks-internal modules)
- mount_and_run moves to new run/ package
  (composes disks.managed_mount + sync.pipeline.check_and_run)
- Delete orchestration.py
